pyproject.toml
[project]
name = "enhanced-linkedin-post-generator"
version = "0.1.0"
description = "LinkedIn Post Generator using Google ADK"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk==0.3.0",
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.0",
    "beautifulsoup4==4.12.3",
    "requests>=2.31.0",
    "lxml==5.1.0",
    "youtube-transcript-api>=1.2.3",
    "textstat==0.7.3",
    "deprecated==1.3.1",
]

[tool.hatch.build.targets.wheel]
packages = ["linkedin_post_agent", "database", "services", "utils"]

